webapi/tct-webstorage-w3c-tests/webstorage/w3c/event_constructor.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>event_session_Constructor</h1>
    <div id="log"></div>
    <script>
        test(function() {
            var t = async_test("storageeventinit test");
            function onStorageEvent(event) {
                t.step(function() {
                    assert_equals(event.type, 'storage');
                    //URL:http://www.w3.org/TR/2013/REC-webstorage-20130730/#dom-storageevent-key
                    //When the object is created, this attribute must be initialized to the empty string.
                    //assert_equals(event.key, null);
                    assert_equals(event.key, '');
                    assert_equals(event.oldValue, null);
                    assert_equals(event.newValue, null);
                    assert_equals(event.url, '');
                    assert_equals(event.storageArea, null);
                });
                t.done();
            }

            window.addEventListener('storage', onStorageEvent, false);
            var event = new StorageEvent('storage');
            window.dispatchEvent(event);
        });
    </script>
 </body>
</html>
